+++ Tests/CoralogixRumTests/CoralogixRumTests.swift
@@ -913,12 +913,6 @@ final class MockSessionReplay: SessionReplayInterface {
     func isInitialized() -> Bool {
         return true
     }
-
-    var maskRectsToReturn: [CGRect]?
-
-    func currentMaskRects() -> [CGRect]? {
-        return maskRectsToReturn
-    }
 }
 
 public final class MockSpan: Span {